+++ include/SmoothConvergence.hpp
@@ -30,20 +30,20 @@ class SmoothConvergence : public FractalColoring {
       return color_map[map_size - 1];
     }
 
-    long double R = Complex::absolute_square(z - z1) /
-                    std::max(1.0L, Complex::absolute_square(z));
-    long double prev_R = Complex::absolute_square(z1 - z2) /
-                         std::max(1.0L, Complex::absolute_square(z1));
+    double R = Complex::absolute_square(z - z1) /
+                    std::max(1.0, Complex::absolute_square(z));
+    double prev_R = Complex::absolute_square(z1 - z2) /
+                         std::max(1.0, Complex::absolute_square(z1));
 
     // Smooth factor
-    long double smooth =
+    double smooth =
         (log2(m_bailout) - log2(prev_R)) / (log2(R) - log2(prev_R));
-    smooth = std::max(0.0L, std::min(1.0L, smooth));
+    smooth = std::max(0.0, std::min(1.0, smooth));
 
-    long double ratio = (map_size - 1) / (long double)(max_iterations);
-    long double value = ratio * iterations;
-    long double prev_value = ratio * (iterations - 1);
-    int index = smooth * value + (1.0L - smooth) * prev_value;
+    double ratio = (map_size - 1) / (double)(max_iterations);
+    double value = ratio * iterations;
+    double prev_value = ratio * (iterations - 1);
+    int index = smooth * value + (1.0 - smooth) * prev_value;
     index = std::max(0, std::min(map_size - 1, index));
     return color_map[index];
   }
