+++ src/Complex.cpp
@@ -6,10 +6,10 @@
 
 namespace PekiProc {
 
-const Complex Complex::ZERO(0.0L, 0.0L);
-const Complex Complex::ONE(1.0L, 0.0L);
+const Complex Complex::ZERO(0.0, 0.0);
+const Complex Complex::ONE(1.0, 0.0);
 
-Complex::Complex(long double real_v, long double imaginary_v)
+Complex::Complex(double real_v, double imaginary_v)
     : real(real_v), imaginary(imaginary_v) {}
 
 Complex::Complex(const Complex& toCopy) {
@@ -45,27 +45,27 @@ Complex Complex::conjugate(const Complex& c) {
   return Complex(c.real, -c.imaginary);
 }
 
-long double Complex::absolute(const Complex& c) {
+double Complex::absolute(const Complex& c) {
   return sqrt(Complex::absolute_square(c));
 }
 
-long double Complex::absolute_square(const Complex& c) {
+double Complex::absolute_square(const Complex& c) {
   return c.real * c.real + c.imaginary * c.imaginary;
 }
 
-long double Complex::getReal() const noexcept {
+double Complex::getReal() const noexcept {
   return real;
 }
 
-long double Complex::getImaginary() const noexcept {
+double Complex::getImaginary() const noexcept {
   return imaginary;
 }
 
-void Complex::setReal(long double value) {
+void Complex::setReal(double value) {
   real = value;
 }
 
-void Complex::setImaginary(long double value) {
+void Complex::setImaginary(double value) {
   imaginary = value;
 }
 
@@ -109,17 +109,17 @@ Complex& Complex::operator-=(const Complex& c) {
 }
 
 Complex Complex::operator*(const Complex& c) {
-  long double real_result = this->real * c.real - this->imaginary * c.imaginary;
-  long double imaginary_result =
+  double real_result = this->real * c.real - this->imaginary * c.imaginary;
+  double imaginary_result =
       this->real * c.imaginary + this->imaginary * c.real;
   return Complex(real_result, imaginary_result);
 }
 
-Complex Complex::operator*(const long double a) {
+Complex Complex::operator*(const double a) {
   return Complex(a * this->real, a * this->imaginary);
 }
 
-Complex& Complex::operator*=(const long double a) {
+Complex& Complex::operator*=(const double a) {
   return *this = *this * a;
 }
 
@@ -128,8 +128,8 @@ Complex& Complex::operator*=(const Complex& c) {
 }
 
 Complex Complex::operator/(const Complex& c) {
-  bool realCompare = CompareDoubles::isEqual(c.real, 0.0L);
-  bool imagCompare = CompareDoubles::isEqual(c.imaginary, 0.0L);
+  bool realCompare = CompareDoubles::isEqual(c.real, 0.0);
+  bool imagCompare = CompareDoubles::isEqual(c.imaginary, 0.0);
   if (realCompare && imagCompare) {
     throw std::invalid_argument("Cannot divide by 0");
   } else if (realCompare && !imagCompare) {
@@ -137,20 +137,20 @@ Complex Complex::operator/(const Complex& c) {
   } else if (!realCompare && imagCompare) {
     return *this / c.real;
   }
-  long double real_result = this->real * c.real + this->imaginary * c.imaginary;
-  long double imag_result = this->imaginary * c.real - this->real * c.imaginary;
-  long double abs_square = Complex::absolute_square(c);
+  double real_result = this->real * c.real + this->imaginary * c.imaginary;
+  double imag_result = this->imaginary * c.real - this->real * c.imaginary;
+  double abs_square = Complex::absolute_square(c);
   return Complex(real_result / abs_square, imag_result / abs_square);
 }
 
-Complex Complex::operator/(const long double a) {
-  if (CompareDoubles::isEqual(a, 0.0L)) {
+Complex Complex::operator/(const double a) {
+  if (CompareDoubles::isEqual(a, 0.0)) {
     throw std::invalid_argument("Cannot divide by 0");
   }
   return Complex(this->real / a, this->imaginary / a);
 }
 
-Complex& Complex::operator/=(const long double a) {
+Complex& Complex::operator/=(const double a) {
   return *this = *this / a;
 }
